Add industry filter to states display
I'd like to edit the attached Python code to add a new filter option to the sidebar in the "States" display mode. The new drop-down menu should be titled "Select Industry". Like in the "Compare to ENOW" display mode, it should be greyed out by default, but users should be able to select a specific industry after selecting a specific sector.