New table and key facts
FIRST, the y-axis of the plots in the Regions, Counties, and States display modes now have too few major intervals. While the exact number of major intervals still be allowed to vary from plot to plot, I would like there to be about 50% more y axis intervals on the average plot

SECOND, in the Regions, Counties, and States display modes, I would like to add a new expandable section just above the existing “Coastal Geographies in Open ENOW” expandable section. This new section should be titled “View as a Table”. If expanded, it should display a table containing the plot data. The columns should represent each year in the range and there should be a row for each Sector (or Industry) shown in the plot above.

THIRD, in the Regions, Counties, and States display modes, I would like to add more key facts immediately below the plot. Currently, the app displays a single line of text in this spot (e.g., “Approximately 3,690,248 people were employed in the selected sector(s) in 2024.”). I would like to keep that line, but replacing “in the selected sector(s)” depending on user’s selection from the sidebar. For example, if the user has selected “All Marine Sectors”, the text should read something like “Approximately 3,690,248 people were employed in all marine sectors in 2024”; if the user has selected “Living Resources”, the text should read something like “Approximately 80,463 people were employed in the Living Resources Sector in 2024”. A new line of italicized, slightly smaller text text immediately below this should summarize the percent change in that total over the year range shown in the plot. For example, if the user has displayed data for years 2015-2024 and employment increased from 50,000 to 60,000 over that period, the new text should read “This represents a 20% increase since 2015.”. If employment decreased from 50,000 to 30,000 over that period, the new text should read “This represents a 40% decrease since 2015.”